requirements.txt
# Core Ollama Integration
ollama
langchain-ollama

# LangChain Core and Extensions
langchain
langchain-community
langchain-chroma
chromadb

# Multi-modal Processing
# Image Processing
pillow
opencv-python
pytesseract
easyocr

# Audio Processing
librosa
soundfile
openai-whisper

# Document Processing
pypdf

# Web Search and Scraping
duckduckgo-search
beautifulsoup4
requests

# NLP
spacy

# Frontend / UI
gradio

# Utilities
numpy
tiktoken